Create project
Created Intellij IDEA project, imported it into an eclipse project so that DMU have something to open.

Added provided Converter.java, modified to be functional - and better.

Attempted to add OPEX for JavaFX abstracts and it's utilities - IDEA works with it perfectly fine as expected but eclipse shits itself. The workarounds to make eclipse work don't translate between the project worked on in IDEA and the DMU test environment in eclipse so it was ditched. If i want abstracts or utilities from OPEX ill just have to copy the sources into this project directly, it's the only way to ensure comparability for when the project's opened in eclipse.

Fucking eclipse.
Intellij master race.